requirements.txt
scipy
requests
pandas
openpyxl
python-dotenv
scikit-learn
matplotlib